issue 9 - crashing when cancelling AccessTokenForm without having entered an access token.
I added a default check in the main form after the AccessTokenForm dialog. This must be done in the main form (not the AccessTokenForm) because the main form creates the AccessTokenForm dialog and then continues.

This still does not address the possibility that a user has entered an incorrect access token (which may be a brand new issue), it only checks that a user has closed the AccessTokenForm without having entered a token at all.