So the user shouldn't be allowed to go past the automatic AccessTokenForm dialog without having entered a token, or it will crash.
The AccessTokenForm dialog generally will only appear automatically the first time a user starts the program, which is why I thought it would be sufficient to simply exit the application in this case.

It could be looped like this until an access token has been set, but then we have the posibility of the user having no possible way to exit the application (both the 'X' and the 'Close' button will continue the loop, the user cannot continue to the main form). This is shown in this commit.

If we do want to implement it this way, we could:
1 - have the MessageBox on line 37 offer an option to close the program, as this MessageBox will only ever popup on the automatic AccessTokenForm dialog,

or 2 - have another button on the AccessTokenForm to exit the application