Then I think this is the most efficient way to do it.
The grading queue is the only form that will throw an error without an access token, so it is not loaded until an access token has been set, whether automatically or manually in settings. The other forms are okay because they are loaded on a button press, and they already have sufficient error messaging.