On second thought it is seems obviously much better to only open the URL if the database write was successful rather than to close it.
Although this will essentially require a Mongo connection to let you use the checkbox as a link to grade the assignment. I think this is fine because there is already a link in teh same row by clicking the assignment name. If we'd still like the checkbox to work as a link without a Mongo connection it would be easy to implement but I don't think its necessary.